+++ Tests/Functional/Testing/TestingFrameworkTest.php
@@ -6,7 +6,6 @@ namespace OliverKlee\Oelib\Tests\Functional\Testing;
 
 use Nimut\TestingFramework\TestCase\FunctionalTestCase;
 use OliverKlee\Oelib\Authentication\FrontEndLoginManager;
-use OliverKlee\Oelib\System\Typo3Version;
 use OliverKlee\Oelib\Testing\TestingFramework;
 use Psr\Http\Message\ServerRequestInterface;
 use TYPO3\CMS\Core\Cache\Backend\AbstractBackend;
@@ -16,6 +15,7 @@ use TYPO3\CMS\Core\Cache\Backend\TransientMemoryBackend;
 use TYPO3\CMS\Core\Cache\CacheManager;
 use TYPO3\CMS\Core\Context\Context;
 use TYPO3\CMS\Core\Core\Environment;
+use TYPO3\CMS\Core\Information\Typo3Version;
 use TYPO3\CMS\Core\TypoScript\TemplateService;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Frontend\Authentication\FrontendUserAuthentication;
@@ -3111,7 +3111,7 @@ final class TestingFrameworkTest extends FunctionalTestCase
      */
     public function disableCoreCachesSetsAllCoreCachesForVersion9(string $identifier, string $backend): void
     {
-        if (Typo3Version::isAtLeast(10)) {
+        if ((new Typo3Version())->getMajorVersion() >= 10) {
             self::markTestSkipped('This test is specific to TYPO3 9LTS.');
         }
 
@@ -3149,7 +3149,7 @@ final class TestingFrameworkTest extends FunctionalTestCase
      */
     public function disableCoreCachesSetsAllCoreCachesForVersion10(string $identifier, string $backend): void
     {
-        if (Typo3Version::isNotHigherThan(9)) {
+        if ((new Typo3Version())->getMajorVersion() <= 9) {
             self::markTestSkipped('This test is specific to TYPO3 10LTS.');
         }